refactor(supervisor): split child process machinery into leaf module
Move spawn/wait/terminate/liveness helpers from main.rs into child.rs to
bring main.rs back under the 600-production-line maintainability target
flagged by scripts/check-maintainability.sh.
